+++ mobile/android/android-components/components/concept/awesomebar/src/main/java/mozilla/components/concept/awesomebar/AwesomeBar.kt
@@ -94,7 +94,8 @@ interface AwesomeBar {
          */
         enum class Flag {
             BOOKMARK,
-            OPEN_TAB
+            OPEN_TAB,
+            CLIPBOARD
         }
 
         /**